# docs/openapi.yaml
openapi: 3.1.0

info:
  title: Mental Health API
  version: 1.0.0

servers:
  - url: http://127.0.0.1:8000/api

paths:

  /sessions:
    get:
      operationId: listSessions
      summary: List all therapy sessions
      description: Returns all therapy sessions ordered by session date.
      tags:
        - Sessions

      responses:
        '200':
          description: List of sessions.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Session'

    post:
      operationId: scheduleSession
      summary: Schedule a therapy session
      description: Creates a new therapy session.
      tags:
        - Sessions

      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ScheduleSessionRequest'

      responses:
        '201':
          description: Session created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Session'

  /sessions/{id}/complete:
    patch:
      tags:
        - Sessions

      operationId: completeSession
      summary: Complete a therapy session.
      description: Marks a therapy session as completed.

      parameters:
        - $ref: '#/components/parameters/SessionId'

      responses:
        '204':
          $ref: '#/components/responses/NoContent'

  /sessions/{id}/cancel:
    patch:
      tags:
        - Sessions

      operationId: cancelSession
      summary: Cancel a therapy session.
      description: Marks a therapy session as cancelled.

      parameters:
        - $ref: '#/components/parameters/SessionId'

      responses:
        '204':
          $ref: '#/components/responses/NoContent'

  /sessions/{id}/reschedule:
    patch:
      tags:
        - Sessions

      operationId: rescheduleSession
      summary: Reschedule a therapy session.
      description: Updates the scheduled date and time of a therapy session.

      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RescheduleSessionRequest'

      parameters:
        - $ref: '#/components/parameters/SessionId'

      responses:
        '204':
          $ref: '#/components/responses/NoContent'

components:

  responses:

    NoContent:
      description: Operation completed successfully. No response body is returned.

  parameters:
    SessionId:
      name: id
      in: path
      required: true
      description: Session identifier.
      schema:
        type: string
        format: uuid
        examples: ["019f4631-5ea6-7324-afac-0f83b6e335bc"]

  schemas:
    Session:
      type: object

      required:
        - id
        - patient_id
        - therapist_id
        - session_date
        - status

      properties:
        id:
          type: string
          format: uuid
          examples: ["019f4631-5ea6-7324-afac-0f83b6e335bc"]

        patient_id:
          type: string
          format: uuid
          examples: ["019f4631-5ea6-7324-afac-0f83b6e335bc"]

        therapist_id:
          type: string
          format: uuid
          examples: ["019f4631-5ea6-7324-afac-0f83b6e335bc"]

        session_date:
          type: string
          format: date-time
          examples: ["2026-07-09T16:56:00Z"]



        status:
          type: string
          enum:
            - scheduled
            - completed
            - cancelled
          examples: ["completed"]

    ScheduleSessionRequest:
      type: object
      required:
        - patient_id
        - therapist_id
        - session_date

      properties:
        patient_id:
          type: string
          format: uuid
          examples: ["019f4631-5ea6-7324-afac-0f83b6e335bc"]

        therapist_id:
          type: string
          format: uuid
          examples: ["019f4631-5ea6-7324-afac-0f83b6e335bc"]

        session_date:
          type: string
          format: date-time
          examples: ["2026-07-09T16:56:00Z"]


    RescheduleSessionRequest:
      type: object
      required:
        - session_date

      properties:
        session_date:
          type: string
          format: date-time